+++ pyonepassword/op_items/item_types/_item_base.py
