+++ pyonepassword/op_items/item_types/login.py
@@ -6,8 +6,8 @@ from .._new_item import OPNewItemMixin
 from .._op_item_type_registry import op_register_item_type
 from ..item_field_base import OPItemField
 from ..item_section import OPSection
+from ._item_base import OPAbstractItem
 from ._item_descriptor_base import OPAbstractItemDescriptor
-from ._op_items_base import OPAbstractItem
 
 
 class OPNewLoginItemURLException(Exception):
