+++ pyonepassword/op_items/item_types/password.py
@@ -1,7 +1,7 @@
 from .._item_descriptor_registry import op_register_item_descriptor_type
 from .._op_item_type_registry import op_register_item_type
+from ._item_base import OPAbstractItem
 from ._item_descriptor_base import OPAbstractItemDescriptor
-from ._op_items_base import OPAbstractItem
 
 # {
 #     "id": "2dvgl7kk5yjrq3gxwqimp5awve",
