+++ pyonepassword/op_items/item_types/server.py
@@ -2,8 +2,8 @@ from typing import Union
 
 from .._item_descriptor_registry import op_register_item_descriptor_type
 from .._op_item_type_registry import op_register_item_type
+from ._item_base import OPAbstractItem, OPFieldNotFoundException
 from ._item_descriptor_base import OPAbstractItemDescriptor
-from ._op_items_base import OPAbstractItem, OPFieldNotFoundException
 
 
 @op_register_item_descriptor_type
