+++ pyonepassword/pyonepassword.py
@@ -13,7 +13,7 @@ from .account import OPAccountList
 from .op_items._item_list import OPItemList
 from .op_items._new_item import OPNewItemMixin
 from .op_items._op_item_type_registry import OPItemFactory
-from .op_items.item_types._op_items_base import OPAbstractItem
+from .op_items.item_types._item_base import OPAbstractItem
 from .op_items.item_types.generic_item import (
     _OPGenericItem,
     _OPGenericItemRelaxedValidation
